refactor: refactor the datafusion planner
